refactored code so the classes Player, Minion, TileManager take the GamePanel as a parameter in the constructor of the previous classes, so the GamePanel is not reiterated again in other methods in those classes. Successfully added the camera to player and current screen moves with player at center of screen, and renders tiles on the screen when player moves left or right!

+++ src/Main.java
@@ -5,6 +5,7 @@ public class Main extends JFrame{
     public static void main(String[] args) {
         GamePanel gPanel = new GamePanel();
         gPanel.init_Thread();
+        //Position the window at the center of the monitor
         double center_winPosX = gPanel.monitor_Width *.5 - (gPanel.screen_Width*.5);
         double center_winPosY = gPanel.monitor_Height * .5 - (gPanel.screen_Height*.5);
         JFrame frame = new JFrame("sideScroller2D");
@@ -12,7 +13,7 @@ public class Main extends JFrame{
         frame.add(gPanel);
         // Converted the double types to int, because parameters of
         // the method the int type
-        frame.setSize((int)gPanel.screen_Width,(int)gPanel.screen_Height);
+        frame.setSize(gPanel.screen_Width,gPanel.screen_Height);
         frame.setLocation((int)center_winPosX,(int)center_winPosY);
         frame.setVisible(true);
     }
